requirements.txt
tqdm
matplotlib
numpy
pillow
torch
torchvision